Move claimed reward task to RewardEraSeqCreatorTask
