+++ indexer/pipeline.go
@@ -67,7 +67,6 @@ func NewPipeline(cfg *config.Config, cli *client.Client, accountDb store.Account
 		pipeline.NewAsyncStageWithTasks(
 			pipeline.StageParser,
 			NewBlockParserTask(),
-			NewTransactionParserTask(cfg, rewardDb, syncableDb, validatorDb),
 			pipeline.RetryingTask(NewValidatorsParserTask(cfg, cli.Account, rewardDb, syncableDb, validatorDb), isTransient, 1),
 		),
 	)
@@ -82,7 +81,7 @@ func NewPipeline(cfg *config.Config, cli *client.Client, accountDb store.Account
 			pipeline.RetryingTask(NewEventSeqCreatorTask(eventDb), isTransient, maxRetries),
 			pipeline.RetryingTask(NewAccountEraSeqCreatorTask(cfg, accountDb, syncableDb), isTransient, maxRetries),
 			pipeline.RetryingTask(NewTransactionSeqCreatorTask(transactionDb), isTransient, maxRetries),
-			pipeline.RetryingTask(NewRewardEraSeqCreatorTask(cfg, syncableDb), isTransient, maxRetries),
+			pipeline.RetryingTask(NewRewardEraSeqCreatorTask(cfg, rewardDb, syncableDb, validatorDb), isTransient, maxRetries),
 		),
 	)
