+++ indexer/sequencer_tasks.go
@@ -2,7 +2,10 @@ package indexer
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"strconv"
+	"strings"
 	"time"
 
 	"github.com/figment-networks/indexing-engine/pipeline"
@@ -10,7 +13,9 @@ import (
 	"github.com/figment-networks/polkadothub-indexer/metric"
 	"github.com/figment-networks/polkadothub-indexer/model"
 	"github.com/figment-networks/polkadothub-indexer/store"
+	"github.com/figment-networks/polkadothub-indexer/types"
 	"github.com/figment-networks/polkadothub-indexer/utils/logger"
+	"github.com/figment-networks/polkadothub-proxy/grpc/event/eventpb"
 )
 
 const (
@@ -22,6 +27,19 @@ const (
 	AccountEraSeqCreatorTaskName       = "AccountEraSeqCreator"
 	TransactionSeqCreatorTaskName      = "TransactionSeqCreator"
 	RewardEraSeqCreatorTaskName        = "RewardEraSeqCreator"
+
+	eventMethodReward     = "Reward"
+	txMethodPayoutStakers = "payoutStakers"
+	txMethodBatch         = "batch"
+	txMethodBatchAll      = "batchAll"
+	txMethodProxy         = "proxy"
+	txMethodSudo          = "sudo"
+	sectionUtility        = "utility"
+	sectionStaking        = "staking"
+	sectionProxy          = "proxy"
+
+	accountKey = "AccountId"
+	balanceKey = "Balance"
 )
 
 var (
@@ -330,13 +348,22 @@ func (t *transactionSeqCreatorTask) Run(ctx context.Context, p pipeline.Payload)
 }
 
 // NewRewardEraSeqCreatorTask creates rewards
-func NewRewardEraSeqCreatorTask(cfg *config.Config, syncablesDb store.Syncables) *rewardEraSeqCreatorTask {
-	return &rewardEraSeqCreatorTask{cfg, syncablesDb}
+func NewRewardEraSeqCreatorTask(cfg *config.Config, rewardsDb store.Rewards, syncablesDb store.Syncables, validatorDb store.ValidatorEraSeq) *rewardEraSeqCreatorTask {
+	return &rewardEraSeqCreatorTask{
+		cfg: cfg,
+
+		rewardsDb:   rewardsDb,
+		syncablesDb: syncablesDb,
+		validatorDb: validatorDb,
+	}
 }
 
 type rewardEraSeqCreatorTask struct {
-	cfg         *config.Config
+	cfg *config.Config
+
+	rewardsDb   store.Rewards
 	syncablesDb store.Syncables
+	validatorDb store.ValidatorEraSeq
 }
 
 func (t *rewardEraSeqCreatorTask) GetName() string {
@@ -355,21 +382,13 @@ func (t *rewardEraSeqCreatorTask) Run(ctx context.Context, p pipeline.Payload) e
 		return err
 	}
 
-	var eraSeq *model.EraSequence
+	// get unclaimed rewards for validators in current era
 	for stash, validator := range payload.ParsedValidators {
 		data := validator.parsedRewards
 
-		eraSeq = currentEraSeq
-		if data.Era != payload.Syncable.Era {
-			eraSeq, err = t.getEraSeq(data.Era, payload.Syncable)
-			if err != nil {
-				return err
-			}
-		}
-
 		if data.Commission != "" {
 			payload.RewardEraSequences = append(payload.RewardEraSequences, model.RewardEraSeq{
-				EraSequence:           eraSeq,
+				EraSequence:           &currentEraSeq,
 				StashAccount:          stash,
 				ValidatorStashAccount: stash,
 				Amount:                data.Commission,
@@ -380,7 +399,7 @@ func (t *rewardEraSeqCreatorTask) Run(ctx context.Context, p pipeline.Payload) e
 
 		if data.Reward != "" {
 			payload.RewardEraSequences = append(payload.RewardEraSequences, model.RewardEraSeq{
-				EraSequence:           eraSeq,
+				EraSequence:           &currentEraSeq,
 				StashAccount:          stash,
 				ValidatorStashAccount: stash,
 				Amount:                data.Reward,
@@ -391,7 +410,7 @@ func (t *rewardEraSeqCreatorTask) Run(ctx context.Context, p pipeline.Payload) e
 
 		if data.RewardAndCommission != "" {
 			payload.RewardEraSequences = append(payload.RewardEraSequences, model.RewardEraSeq{
-				EraSequence:           eraSeq,
+				EraSequence:           &currentEraSeq,
 				StashAccount:          stash,
 				ValidatorStashAccount: stash,
 				Amount:                data.RewardAndCommission,
@@ -402,7 +421,7 @@ func (t *rewardEraSeqCreatorTask) Run(ctx context.Context, p pipeline.Payload) e
 
 		for _, n := range data.StakerRewards {
 			payload.RewardEraSequences = append(payload.RewardEraSequences, model.RewardEraSeq{
-				EraSequence:           eraSeq,
+				EraSequence:           &currentEraSeq,
 				StashAccount:          n.Stash,
 				ValidatorStashAccount: stash,
 				Amount:                n.Amount,
@@ -412,15 +431,54 @@ func (t *rewardEraSeqCreatorTask) Run(ctx context.Context, p pipeline.Payload) e
 		}
 	}
 
+	// get claimed rewards from staking.payoutStakers txs
+	for _, tx := range payload.RawBlock.GetExtrinsics() {
+		var claims []RewardsClaim
+		var err error
+
+		if !tx.GetIsSuccess() {
+			continue
+		}
+
+		if tx.GetSection() == sectionStaking && tx.GetMethod() == txMethodPayoutStakers {
+			claim, err := getRewardsClaimFromPayoutStakersTx(tx.GetArgs())
+			claim.TxHash = tx.GetHash()
+			if err != nil {
+				return err
+			}
+			claims = append(claims, claim)
+		} else if (tx.GetSection() == sectionUtility && (tx.GetMethod() == txMethodBatch || tx.GetMethod() == txMethodBatchAll)) ||
+			(tx.GetSection() == sectionProxy && tx.GetMethod() == txMethodProxy) {
+			for _, callArg := range tx.GetCallArgs() {
+				if callArg.GetSection() == sectionStaking && callArg.GetMethod() == txMethodPayoutStakers {
+					claim, err := getRewardsClaimFromPayoutStakersTx(callArg.GetValue())
+					if err != nil {
+						return err
+					}
+					claim.TxHash = tx.GetHash()
+					claims = append(claims, claim)
+				}
+			}
+		}
+
+		rewards, rewardclaims, err := t.getRewardsFromEvents(tx.GetExtrinsicIndex(), claims, payload.RawEvents)
+		if err != nil {
+			return err
+		}
+
+		payload.RewardsClaimed = append(payload.RewardsClaimed, rewardclaims...)
+		payload.RewardEraSequences = append(payload.RewardEraSequences, rewards...)
+	}
+
 	return nil
 }
 
-func (t *rewardEraSeqCreatorTask) getEraSeq(era int64, currentSyncable *model.Syncable) (*model.EraSequence, error) {
+func (t *rewardEraSeqCreatorTask) getEraSeq(era int64, currentSyncable *model.Syncable) (model.EraSequence, error) {
 	var firstHeightInEra int64
 	lastSyncableInPrevEra, err := t.syncablesDb.FindLastInEra(era - 1)
 	if err != nil {
 		if err != store.ErrNotFound {
-			return nil, err
+			return model.EraSequence{}, err
 		}
 		firstHeightInEra = t.cfg.FirstBlockHeight
 	} else {
@@ -428,17 +486,186 @@ func (t *rewardEraSeqCreatorTask) getEraSeq(era int64, currentSyncable *model.Sy
 	}
 
 	lastSyncableInEra := currentSyncable
-	if currentSyncable.Era != era {
+	if currentSyncable == nil || currentSyncable.Era != era {
 		lastSyncableInEra, err = t.syncablesDb.FindLastInEra(era)
 		if err != nil {
-			return nil, err
+			return model.EraSequence{}, err
 		}
 	}
 
-	return &model.EraSequence{
+	return model.EraSequence{
 		Era:         era,
 		StartHeight: firstHeightInEra,
 		EndHeight:   lastSyncableInEra.Height + 1,
 		Time:        lastSyncableInEra.Time,
+	}, nil
+}
+
+// func (t *rewardEraSeqCreatorTask) getEraSeq(era int64) (model.EraSequence, error) {
+// 	var firstHeightInEra int64
+// 	lastSyncableInPrevEra, err := t.syncablesDb.FindLastInEra(era - 1)
+// 	if err != nil {
+// 		if err != store.ErrNotFound {
+// 			return model.EraSequence{}, err
+// 		}
+// 		firstHeightInEra = t.cfg.FirstBlockHeight
+// 	} else {
+// 		firstHeightInEra = lastSyncableInPrevEra.Height + 1
+// 	}
+
+// 	lastSyncableInEra, err := t.syncablesDb.FindLastInEra(era)
+// 	if err != nil {
+// 		return model.EraSequence{}, err
+// 	}
+
+// 	return model.EraSequence{
+// 		Era:         era,
+// 		StartHeight: firstHeightInEra,
+// 		EndHeight:   lastSyncableInEra.Height,
+// 		Time:        lastSyncableInEra.Time,
+// 	}, nil
+// }
+
+// getRewardsFromEvents rreturns claims if rewards exist already in db, and returns new reward seqs if reward seqs don't exist in db
+func (t *rewardEraSeqCreatorTask) getRewardsFromEvents(txIdx int64, claims []RewardsClaim, events []*eventpb.Event) ([]model.RewardEraSeq, []RewardsClaim, error) {
+	var rewards []model.RewardEraSeq
+	var rewardClaims []RewardsClaim
+
+	if len(events) == 0 && len(claims) != 0 {
+		return rewards, rewardClaims, fmt.Errorf("[getRewardsFromEvents] expected events to not be empty")
+	}
+
+	var idx int
+	var nextVald string
+	for i, claim := range claims {
+		if i < len(claims)-1 {
+			nextVald = claims[i+1].ValidatorStash
+		} else {
+			nextVald = ""
+		}
+
+		eraSeq, err := t.getEraSeq(claim.Era, nil)
+		if err != nil {
+			return rewards, rewardClaims, err
+		}
+
+		extractedRewards, ranged, err := t.extractRewardsForClaimFromEvents(claim, eraSeq, txIdx, nextVald, events[idx:])
+		if err != nil {
+			return rewards, rewardClaims, err
+		}
+		idx += ranged
+
+		count, err := t.rewardsDb.GetCount(claim.ValidatorStash, claim.Era)
+		if err != nil {
+			return rewards, rewardClaims, err
+		}
+
+		if count == 0 {
+			rewards = append(rewards, extractedRewards...)
+			continue
+		}
+
+		// if already exists in db, then claim was added to RewardsClaimed, so don't add parsedRewards
+		// instead check that count in db matches parsedRewards. Count may be +1 more since unclaimed validator rewards
+		// can be split into commission and reward
+		if int(count) != len(extractedRewards) && int(count) != len(extractedRewards)+1 {
+			return rewards, rewardClaims, fmt.Errorf("Expected unclaimed rewards to match number of rewards from claim %v; got: %v want: %v (~-1)", claim, len(extractedRewards), count)
+		} else {
+			rewardClaims = append(rewardClaims, claim)
+		}
+	}
+
+	return rewards, rewardClaims, nil
+}
+
+func (t *rewardEraSeqCreatorTask) extractRewardsForClaimFromEvents(claim RewardsClaim, eraSeq model.EraSequence, txIdx int64, nextVald string, events []*eventpb.Event) ([]model.RewardEraSeq, int, error) {
+	var rewards []model.RewardEraSeq
+	var foundCurrVald bool
+
+	for i, event := range events {
+		if event.GetExtrinsicIndex() != txIdx || event.GetMethod() != eventMethodReward || event.GetSection() != sectionStaking {
+			continue
+		}
+
+		stash, amount, err := t.getStashAndAmountFromData(event)
+		if err != nil {
+			return rewards, i, err
+		}
+
+		if foundCurrVald && stash == nextVald {
+			return rewards, i, nil
+		}
+
+		if stash == claim.ValidatorStash {
+			foundCurrVald = true
+		}
+
+		reward := model.RewardEraSeq{
+			Kind:                  model.RewardReward,
+			EraSequence:           &eraSeq,
+			StashAccount:          stash,
+			Amount:                amount.String(),
+			ValidatorStashAccount: claim.ValidatorStash,
+			Claimed:               true,
+			TxHash:                claim.TxHash,
+		}
+
+		if stash == claim.ValidatorStash {
+			reward.Kind = model.RewardCommissionAndReward
+		}
+
+		rewards = append(rewards, reward)
+	}
+
+	if nextVald != "" {
+		return rewards, 0, fmt.Errorf("Could not extract rewards from events")
+	}
+
+	return rewards, 0, nil
+}
+
+func (t *rewardEraSeqCreatorTask) getStashAndAmountFromData(event *eventpb.Event) (stash string, amount types.Quantity, err error) {
+	for _, d := range event.GetData() {
+		if d.GetName() == accountKey {
+			stash = d.Value
+		} else if d.GetName() == balanceKey {
+			amount, err = types.NewQuantityFromString(d.Value)
+		}
+	}
+	if stash == "" {
+		err = errUnexpectedEventDataFormat
+	}
+	return
+}
+
+func getRewardsClaimFromPayoutStakersTx(args string) (RewardsClaim, error) {
+	var data []string
+
+	err := json.Unmarshal([]byte(args), &data)
+	if err == nil {
+		return getStashAndEraFromPayoutArgs(data)
+	}
+
+	parts := strings.Split(args, ",")
+	if len(parts) != 2 {
+		return RewardsClaim{}, errUnexpectedEventDataFormat
+	}
+
+	return getStashAndEraFromPayoutArgs(parts)
+}
+
+func getStashAndEraFromPayoutArgs(data []string) (RewardsClaim, error) {
+	if len(data) < 2 {
+		return RewardsClaim{}, errUnexpectedTxDataFormat
+	}
+
+	era, err := strconv.ParseInt(data[1], 10, 64)
+	if err != nil {
+		return RewardsClaim{}, err
+	}
+
+	return RewardsClaim{
+		Era:            era,
+		ValidatorStash: data[0],
 	}, nil
 }
