+++ indexer/syncer_tasks.go
@@ -3,6 +3,7 @@ package indexer
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"time"
 
 	"github.com/figment-networks/polkadothub-indexer/metric"
